+++ browser-test/clear_data.py
@@ -23,12 +23,12 @@ if OPENWISP_RADIUS_PATH == '':
 
 sys.path.insert(0, os.path.join(OPENWISP_RADIUS_PATH, 'tests'))
 sys.argv.insert(1, 'browser-test')
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'openwisp2.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'immunity2.settings')
 try:
     django.setup()
 except ImportError:
     print(
-        'OpenWISP RADIUS is not installed or python '
+        'Immunity RADIUS is not installed or python '
         'virtual environment is not activated correctly',
         file=sys.stderr,
     )
@@ -38,8 +38,8 @@ except ImportError:
 from django.contrib.auth import get_user_model
 
 User = get_user_model()
-RadiusAccounting = swapper.load_model('openwisp_radius', 'RadiusAccounting')
-Organization = swapper.load_model('openwisp_users', 'Organization')
+RadiusAccounting = swapper.load_model('immunity_radius', 'RadiusAccounting')
+Organization = swapper.load_model('immunity_users', 'Organization')
 
 test_data = load_test_data()