+++ server/controllers/payments-controller.js
@@ -3,7 +3,7 @@ import merge from "deepmerge";
 import config from "../config.json";
 import defaultConfig from "../utils/default-config";
 import {logResponseError} from "../utils/logger";
-import reverse from "../utils/openwisp-urls";
+import reverse from "../utils/immunity-urls";
 import getSlug from "../utils/get-slug";
 
 const payments = (req, res) => {